refactor: Slim down FolioReaderCenter.swift and extract extensions (P3-8)
